Speaker Feedback: Add "inappropriate" comment status for feedback that is not spam but should not be sent to the speaker (#433)
This shoehorns in a custom comment status, despite WP Core's best efforts to prevent it. Feedback comments can be marked as inappropriate via bulk editing in the list table. A new list table view shows just feedback marked with this new status, along with a total count.

It also adds a list table view for feedbacks that the speaker has marked as "helpful".

Fixes #414