Speaker Feedback: Consistently use sessions to determine start/end times of the event (#451)
Currently we are using different values in different places to determine when the WordCamp event as a whole starts and ends. Some places we are using the start time of the earliest session as a start time, some places we are using `Start Date (YYYY-MM-DD)` from the WordCamp post. For the ending time, we are mostly extrapolating from whatever we're using as the start time.

This creates new functions for getting a timestamp for the earliest session and a timestamp for the end of the latest session, which I think are the most relevant numbers to use in the context of this plugin, which is all about the sessions.

This also makes it easier for us to test the plugin on a production site. We can use a test site that doesn't have an associated WordCamp post, as long as the session times are set up correctly.