WCPT: Use separate Code of Conduct page stub for online events (#531)
Online only events use slightly different wording in the default Code of Conduct. This uses a separate stub with that wording when creating a new site if the "Virtual event only" box is checked for the WordCamp.

This also removes the get_code_of_conduct method from the `WordCamp_New_Site` class in favor of putting the content directly into the stub files.

Additionally, this converts the CoC content from a complex ordered list to simpler heading and paragraph blocks. This makes it easier for organizers to manage and customize the content and also eliminates some issues with extra line breaks.

Fixes #497