Helpers: Remove unnecessary `URL` query.
The `set_site_id_post_meta.php` script added a `_site_id` value to all of the `wordcamp` posts that have corresponding sites, so there's no longer a need to search by `URL` too. There are still about `650` posts that don't have the `_site_id` value, but that's because they don't have a site either (or it's external, etc).