Super Cache: Remove inactive Tagregator plugin.
It depends on https://github.com/Automattic/wp-super-cache/pull/121/ in order to work, which hasn't been merged. We used to run that branch of the plugin on production, but stopped doing that several years ago. At that point, this plugin became inactive, and it's no longer doing anything.

We have some caching at the Nginx layer in order to prevent overloading the site if the the Tagregator endpoint gets too much traffic, so this is no longer needed anyway.