WCPT: Create absolute links to support `city/year` URLs.
Previously, the `Call for Sponsors` post would be created with a link to `/call-for-sponsors`. On `year.city` URLs, that would correctly point to `{year}.{city}.wordcamp.org/call-for-sponsors`. On `{city}.wordcamp.org/{year}/` URLs, though, it points to `{city}.wordcamp.org/call-for-sponsors/`, and ends up redirecting to the home page of the canonical year.