Sunrise: Disable broken `india/2020` redirect.
It was setup in d5650a7d, but didn't have any effect, because `site_redirects()` doesn't support parsing `city/year` URL formats yet. The redirect itself still happened, but because of `get_canonical_year_url()`.

That parsing support needs to be added, but currently the Dehydrated endpoint is broken, because it's outputting an entry for `india.wordcamp.org/2020 www.india.wordcamp.org/2020`, which aren't valid FQDNs.

To fix that, `WordCamp_Lets_Encrypt_Helper::get_domains()` needs to add parsing support as well, but this is a quick workaround to disable error notifications.