Central: Fix white slide flash in homepage slideshow.
The display: none set by `hide` was not being removed until the end of the fadeIn animation, causing the appearance of a white slide during the animation. Instead, set and animate opacity property directly to avoid display issues. The images are absolutely positioned in a stack, so leaving them in the page flow is okay. This also updates the z-index of the active slide to ensure it's always on top.