Blocks: Improve date display for schedule & sessions across timezones (#696)
* Blocks: Use format when calculating values for Schedule block

`wp.date.format` will respect the client's timezone, and show the date in the viewer's timezone. The current `dateI18n` approach always uses the site timezone, which can be confusing for regional and online events.

* Prefix the grid rules with the date

In some cases, if a session spans later than midnight and ends at the same time as an earlier timeslot, it can cause the sessions to overlap because the grid slot is already defined. Prepending the date fixes this, since separate days have separate grid systems.

* Add timezone to the sessions in Session and Speakers blocks

This is mostly a workaround, instead of showing the user's time, but it's an improvement

* Display row-heading date with timezone