Blocks: Fix undefined errors in schedule block front end.
These two things were previously added to the page by the live-schedule & live-posts blocks, but those scripts are not enqueued in block themes unless the blocks are used in the content. This was causing two JS errors when trying to view the schedule— `WordCampBlocks` is undefined, and `wp.sanitize` is undefined. The fix here is to create `WordCampBlocks` if needed, and add `wp-sanitize` to the frontend dependencies.