Jetpack Tweaks: Filter the `is_frontend` check to prevent broken contact forms.
Something in WP Super Cache is caching the "submit a form" link, which happens when Jetpack thinks it's not the frontend of the site. While waiting for this to be fixed in Jetpack, this workaround updates the logic to almost-always say it is "frontend" and render the form.

See https://github.com/Automattic/jetpack/issues/22410.