Camptix: Remove legacy lazy-loading for attendee avatars (#730)
Browser-based lazy loading for images has been built into WP since 5.4. We no longer need to use JavaScript and various jQuery scripts to accomplish this for the attendee avatars output by the Attendees shortcode.

This removes all the scripts and markup for our old homegrown lazy-loading scheme and just outputs the avatars via WP's get_avatar function. Core takes care of the rest.

This also changes the expiration for the cached attendee list markup to refresh every 3 hours instead of 24, as there have been a few complaints about that in the past when people wanted to see themselves show up on the attendees page.

Refs #374