Spam Prevention: Lower threshold to more realistic value
This accounts for folks who submit a form, get an error, then make a quick change and re-submit. All of the other fields were prefilled, so it could be reasonably done in just a few seconds. Bots will still be faster than that.

This also makes it configurable, for forms that need to be more/less strict.