Blocks: Check for post context values before using data (#851)
If used in the post editor, these are valid and reflect the current post. If used in the template editor, these are null, and the block should display a placeholder value. This will display the placeholder value in the template editor regardless of sessions (published or not) on the site.

Fixes #829, fixes #832